re-organised tests
