+++ DfE.GIAP.All/DfE.GIAP.Web/Views/Home/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model LandingViewModel
+﻿@model HomeViewModel
 @{
     var title = string.IsNullOrEmpty(Model.LandingResponse?.Title) ? "Get information about pupils" : Model.LandingResponse.Title;
     ViewData["Title"] = title;
