--- DfE.GIAP.All/DfE.GIAP.Web/Views/Landing/Index.cshtml
@@ -1,64 +1,0 @@
-﻿@model LandingViewModel
-@{
-    var title = string.IsNullOrEmpty(Model.LandingResponse?.Title) ? "Get information about pupils" : Model.LandingResponse.Title;
-    ViewData["Title"] = title;
-}
-
-<div class="govuk-grid-row">
-    <div class="govuk-grid-column-three-quarters">
-        <h1 class="govuk-heading-xl" id="landing--heading--text">
-            @Html.Raw(title)
-        </h1>
-        <p class="govuk-body">
-            @Html.Raw(Model.LandingResponse?.Body)
-        </p>
-
-        <form asp-controller="Landing" asp-action="Index" method="post">
-            <button class="govuk-button govuk-button--start" id="landing--continue--btn" type="submit">
-                Search @Html.Raw(!User.IsOrganisationEstablishmentWithFurtherEducation() ? "national pupil database" : "further education")
-                <svg class="govuk-button__start-icon" xmlns="http://www.w3.org/2000/svg" width="17.5" height="19" viewBox="0 0 33 40" aria-hidden="true" focusable="false">
-                    <path fill="currentColor" d="M0 0h13l20 20-20 20H0l20-20z" />
-                </svg>
-            </button>
-        </form>
-
-        @if (!Model.PlannedMaintenanceResponse.IsEmpty())
-        {
-            <div id="landing--maintenance--container">
-                <h2 class="govuk-heading-m govuk-!-margin-top-1 govuk-!-text-break-word" id="landing--plannedwork-heading--text">
-                    @Model.PlannedMaintenanceResponse.Title
-                </h2>
-                <p class="govuk-body" id="landing--plannedwork-content--text">
-                    @Html.Raw(Model.PlannedMaintenanceResponse.Body)
-                </p>
-            </div>
-        }
-        @if (!Model.PublicationScheduleResponse.IsEmpty())
-        {
-            <div id="landing--publication--container">
-                <h2 class="govuk-heading-m govuk-!-margin-top-1 govuk-!-text-break-word" id="landing--release-heading--text">
-                    @Html.Raw(Model.PublicationScheduleResponse?.Title)
-                </h2>
-                <p class="govuk-body" id="landing--release-description--text">
-                    @Html.Raw(Model.PublicationScheduleResponse?.Body)
-                </p>
-            </div>
-        }
-        @if (!Model.FAQResponse.IsEmpty())
-        {
-            <div id="landing--faq--container">
-                <h2 class="govuk-heading-m govuk-!-margin-top-1 govuk-!-text-break-word" id="landing--faq-heading--text">
-                    @Html.Raw(Model.FAQResponse?.Title)
-                </h2>
-                <p class="govuk-body webapp-formatted-text" id="landing--faq-description--text">
-                    @Html.Raw(Model.FAQResponse?.Body)
-                </p>
-            </div>
-        }
-
-    </div>
-
-    <div class="govuk-grid-column-one-quarter">
-        <partial name="Layout/_LayoutSideMenuPopularActivities" />
-    </div>
-</div>
